Implemented a group_by=id clause
What?

This change allows client to perform a group_by=ID. Since ID's are
unique, that means all individual lot entries are returned. No
grouping or aggregation is performed.

Implementation?

A convenience utility is added at the Service layer. Since the select
and the group_by clause are made to always the same, to get all the
properties of a lot, the client needs to specify all column_names like
group_by=ID,MATERIAL,QUANTITY,SHIPMENT,STORAGE otherwise all those
attributes would be null by default. But since group_by having ID
guarantees that all tuples will be returned, we can do one better.
When group_by=ID is present in the request, the service adds all
the remaining fields to the select and group_by clause. This makes
it convenient for the client to get all the results and not have to
specify an entire list of attributes.

Other changes:

Implemented annotation for specifying custom join types when using the
Criteria API.

What?

Since LotAggregationService uses Criteria API to create dynamic queries
at runtime, the annotations specified on the entity attributes are ignored
when queries are generated. By default, the joins between the child
entities are created an inner joins. This creates a bug because when a
child entity is null, the inner join results in 0 tuples. Hence, no items
are returned.

Implementation?

A new annotation is created named `CriteriaJoin` that lets the developer
specify the JoinType relationship. This is used in the AggregationService
to retrieve the type of join that must be created between the parent entity
and the child attribute.

Also includes:
- Renaming of class names for consistency.
- Refactoring of classes to have GroupBy and Select clause separately.
- Empty line cleanup
- Modified the PG containers to log the SQL statement being run on them. This provides a greater insight to performance and helps with debugging
- Removed the deployment item from template checklist since auto-deploy is enabled.