Added Mutation Testing
What?

Mutation testing is a way to test the effectiveness of the tests.
This change introduces the pi-test plugin that performs mutation
testing when MUTATION_COVERAGE flag is set to .
Caution: Mutation testing is intensive and takes about 1 hour to
complete in our case. It is not recommended to be run during a normal
build process.

What else?

The change introduces profiles in the POM file to do the following:
- Only run Mutation testing when flag MUTATION_COVERAGE is set to true
- Skip the tests when SKIP_TESTS flag is set to true in the  file
- Export the env variables in the Pipelines so that mutation testing is not
run but unit tests are performed as part of the build