Modified the Shipment and Invoice to have duplicate children (#183)
What?

Previously, Invoices and Shipments enforced that a duplicate child (an invoiceItem or a lot, respectively)
cannot be added to the them. But this creates a problem in case of the procurement API. Also, there may be
some edge cases where multiple duplicate children are needed. So, this change removes that constraint.

Issue: https://github.com/northstacksoftware/brewcraft-backend/issues/180