Removed NON_NULL from JSON serialization.
What?

It's confusing for the developers to use the API when null valued properties are missing altogether
from the response. While it's efficient to keep this setting, it's not worth the confusion and the
lost productivity. Hence, removing it altogether seems the best option.

Other:
I also removed the Include.ALWAYS from the CategoryDto.java class for cleanup