Implemented Predefined GroupBy collection to Inventory API (#194)
* Implemented Predefined GroupBy collection to Inventory API

What?

The LotAggregationService (aka Inventory service) gave complete control to the client
on what values it wants to pass in the group_by clause. But that made it complicated and
created confusion.

Also, some of the parameters were supposed to be used together and hence some logic and checks
needed to be put in place to include the missing parameters which was code-smell. This change
fixes that.

How?

This change introduce AggregationFieldCollection classes which are accepted by the RestController.
These enums are basically a predefined collection of AggregationFields that belong together and
the ones that client can use more easily.

* Fixed constructors

* Fixed the merge conflict error

* Added non null annotation to Inventory lots as null values don't make sense in their case.

* Code review changes

* Added comment

* Added option to group by lotnumber and invoiceItem both