BugFix for issue on PUT procurements
What?

On a PUT call to procurements endpoints, the invoices are updated first and then the shipment (and material-lots).
In case, when a PUT call removes an existing InvoiceItem, it throws an error because an existing material-lot (which
will be eventually be removed) still refers the old invoice-item (and it violates foreign key constraint).

How?

This change makes the material-lot refer a null value when invoice-item is deleted. This way the invoices are updated
(thereby invoiceItem is removed). After that, material-lot is also removed.