Fixed the CriteriaJoiner bug
What?

The CriteriaJoiner always performed the Root.get(name) operation on the last node of the path.
The get method creates an implicit inner join by default. Therefore, it ignored what JoinType was
specified in the @CriteriaJoin annotation altogether.
This is fixed in this changeset

How?

Instead of doing a simple get method call on the root entity, there is a conditional put in place
If the field in question qualifies as a basic entity, then a call to 'get' is made because in that
case it doesn't make sense to do a join (as it's basic sql column). On the contrary, when entity is
compound, it always creates a join and the type of the join is fetched from the CriteriaJoin annotation