BugFix: Implemented ordering for the Invoice and Shipment children
What?

Currently, when items are fetched by the parent, no order is specified and therefore
the order cannot be guaranteed.
This creates a problem for ProcurementItems since they are created at runtime from the fetched
InvoiceItems and MaterialLots. Since their order is not fixed, it can result in unmatching
procurementItems to be created.

How?

This changeset introduces an index field on the InvoiceItem and MaterialLot that is used to
order the children