Inventory Quantity Fix
What?

We persist all the quantities for materials lots in system units and perform
the conversion back and forth to show them in the display unit. But this creates an
error when we retrieve the inventory because during the aggregation we remove the
display unit information and show it in system unit.

How?

This change uses the baseDisplayUnit associated with the Material to render the
inventory system quantity into more displayable value.

Other changes:

- A lot of unit tests used KG as unit for test data. But since systemUnit for Mass
  is also KG, it had some pitfalls because a lot of entities perform converstions
  to and from the system unit. In case both these units are KG and we missed if
  there was a bug in this situation. I refactored the tests to use GRAMS instead.

- Refactored the code to use SupportedUnit class instead of the Units class because