Tenant persist a random UUID when NULL is specified (#248)
* BugFix: SecureRandom freezes unit tests on Jenkins

What?

In a previous change, we replaced the usage of java.util.Random with
SecureRandom class for generating tenant db-user password. SecureRandom
is extremely slow when running as part of the CICD pipeline and caused
the builds to take longer than an hour.

Why?

SecureRandom uses /dev/random to generate values. /dev/random is slow
if the host machine needs to wait to build entropy. Since we're spinning
up a headless docker container, they neither have entropy nor have any
activity to build it. This causes huge detail to get a random value for
unit tests.

How?
Out of multiple ways to fix it, I went with refactoring the code to inject
an instance of the SecureRandom class so that we can pass a mocked one at
the unit tests. This also reduces the flakiness of the test since we always
get the same password output now.

* Tenant persist a random UUID when NULL is specified

* Removed unused annotations