Inventory Quantity Fix (#253)
* BugFix: SecureRandom freezes unit tests on Jenkins

What?

In a previous change, we replaced the usage of java.util.Random with
SecureRandom class for generating tenant db-user password. SecureRandom
is extremely slow when running as part of the CICD pipeline and caused
the builds to take longer than an hour.

Why?

SecureRandom uses /dev/random to generate values. /dev/random is slow
if the host machine needs to wait to build entropy. Since we're spinning
up a headless docker container, they neither have entropy nor have any
activity to build it. This causes huge detail to get a random value for
unit tests.

How?
Out of multiple ways to fix it, I went with refactoring the code to inject
an instance of the SecureRandom class so that we can pass a mocked one at
the unit tests. This also reduces the flakiness of the test since we always
get the same password output now.

* Inventory Quantity Fix

What?

We persist all the quantities for materials lots in system units and perform
the conversion back and forth to show them in the display unit. But this creates an
error when we retrieve the inventory because during the aggregation we remove the
display unit information and show it in system unit.

How?

This change uses the baseDisplayUnit associated with the Material to render the
inventory system quantity into more displayable value.

Other changes:

- A lot of unit tests used KG as unit for test data. But since systemUnit for Mass
  is also KG, it had some pitfalls because a lot of entities perform converstions
  to and from the system unit. In case both these units are KG and we missed if
  there was a bug in this situation. I refactored the tests to use GRAMS instead.

- Refactored the code to use SupportedUnit class instead of the Units class because