BugFix: Unique constraint violation when updating Invoice/Shipment
What?

The INVOICE_ITEM table has a constraint to make sure there is only one item
for a given index value for a invoice. This creates a problem when entity is
updated because invoiceItems are a managed collection. When new items are added
and the old is removed, the new one will have the same index as the old (to-be-removed)
invoiceItem.

This causes a SQL exception because when invoice is persisted, the hibernate
performs an insert of the new entities, then updates the existing and finally
removes the orphans.

The procedure fails on the insert step because at that point there is already
the old invoiceItem present with the index value.

How?

This change makes the constaint deferrable. So instead of enforcing the uniqueness
at the end of each statement, now the constraint is enforced after the whole
transaction is complete.