BugFix: Null currency for the tax value returned empty results
Hibernate by default performed an inner join on the currency table. When the value is
null in the invoice table, it results in empty result set. Using lazy fetch means
that the join is not longer performed in the select clause for invoice and the currency value
is fetched in a separate query.

PS: This could cause side-effects elsewhere so we need to keep an eye out